+++ drop-boxes/register-attachments-dropbox/register-attachment-dropbox.py
@@ -84,10 +84,10 @@ def process(transaction):
 	sa = transaction.getSampleForUpdate(sampleID)
 	space = sa.getSpace()
 	if not attachmentReady:
-		expID = '/' + space + '/' + project + '/'+ project+'_INFO'
-		exp = transaction.getExperimentForUpdate(expID)
-		if not exp:
-			exp = transaction.createNewExperiment(expID, "Q_PROJECT_DETAILS")
+		infoSampleID = "/"+space+"/"+code
+		sa = transaction.getSampleForUpdate(infoSampleID)
+	if not sa:
+		exp = transaction.createNewExperiment('/' + space + '/' + project + '/'+ project+'_INFO', "Q_PROJECT_DETAILS")
 		sa = transaction.createNewSample('/' + space + '/'+ code, "Q_ATTACHMENT_SAMPLE")
 		sa.setExperiment(exp)
 	info = None