+++ drop-boxes/register-omero-metadata/image_registration_process.py
@@ -7,30 +7,47 @@ import csv
 from subprocess import Popen, PIPE
 
 barcode_pattern = re.compile('Q[a-zA-Z0-9]{4}[0-9]{3}[A-Z][a-zA-Z0-9]')
+conda_home_path = "/home/qeana10/miniconda2/"
+omero_lib_path = "/home/qeana10/openbis/servers/core-plugins/QBIC/1/dss/drop-boxes/register-omero-metadata/OMERO.py-5.4.10-ice36-b105"
+etl_home_path = "/home/qeana10/openbis/servers/core-plugins/QBIC/1/dss/drop-boxes/register-omero-metadata/"
+
 
 class ImageRegistrationProcess:
 
-    def __init__(self, transaction, env_name="omero_env_0", project_code="", sample_code=""):
+    def __init__(self, transaction, env_name="omero_env_0", project_code="", sample_code="", conda_path=None, omero_path=None, etl_path=None):
 
         self._transaction = transaction
         self._incoming_file_name = transaction.getIncoming().getName()
+        self._search_service = transaction.getSearchService()
 
         self._project_code = project_code
         self._sample_code = sample_code
 
+        ### set exec. env
+        self._conda_path = conda_home_path
+        if not conda_path is None:
+            self._conda_path = conda_path
+
+        self._omero_path = omero_lib_path
+        if not omero_path is None:
+            self._omero_path = omero_path
+
+        self._etl_path= etl_home_path
+        if not etl_path is None:
+            self._etl_path = etl_path
+
         self._init_cmd_list = []
-        self._init_cmd_list.append('eval "$(/home/qeana10/miniconda2/bin/conda shell.bash hook)"')
+        self._init_cmd_list.append('eval "$(' + self._conda_path + 'bin/conda shell.bash hook)"')
         self._init_cmd_list.append('conda activate ' + env_name)
 
-        self._init_cmd_list.append('export OMERO_PREFIX=/home/qeana10/openbis/servers/core-plugins/QBIC/1/dss/drop-boxes/register-omero-metadata/OMERO.py-5.4.10-ice36-b105')
+        self._init_cmd_list.append('export OMERO_PREFIX=' + self._omero_path)
         self._init_cmd_list.append('export PYTHONPATH=$PYTHONPATH:$OMERO_PREFIX/lib/python')
         
-        #now use the omero-importer app packaged in the conda env
-        #self._init_cmd_list.append('export PATH=$PATH:/home/qeana10/openbis/servers/core-plugins/QBIC/1/dss/drop-boxes/register-omero-metadata/OMERO.server-5.4.10-ice36-b105/bin')
-        self._init_cmd_list.append('export PATH=$PATH:/home/qeana10/miniconda2/envs/' + env_name + '/bin')
+        # now use the omero-importer app packaged in the conda env
+        self._init_cmd_list.append('export PATH=$PATH:' + self._conda_path + 'envs/' + env_name + '/bin')
 
-        #move to the dir where backendinterface.py lives
-        self._init_cmd_list.append('cd /home/qeana10/openbis/servers/core-plugins/QBIC/1/dss/drop-boxes/register-omero-metadata/')
+        # move to the dir where backendinterface.py lives for exec.
+        self._init_cmd_list.append('cd ' + self._etl_path)
 
     def fetchOpenBisSampleCode(self):
         found = barcode_pattern.findall(self._incoming_file_name)
@@ -43,7 +60,17 @@ class ImageRegistrationProcess:
             raise SampleCodeError(self._sample_code, "The sample code seems to be invalid, the checksum could not be confirmed.")
 
         return self._project_code, self._sample_code
-    
+
+    def searchOpenBisSample(self, sample_code):
+        # find specific sample
+        sc = SearchCriteria()
+        sc.addMatchClause(SearchCriteria.MatchClause.createAttributeMatch(SearchCriteria.MatchClauseAttribute.CODE, sample_code))
+        foundSamples = self._search_service.searchForSamples(sc)
+        if len(foundSamples) == 0:
+            raise SampleNotFoundError(sample_code, "Sample could not be found in openBIS.")
+        sample = foundSamples[0]
+        return sample
+
     def _isValidSampleCode(self, sample_code):
         try:
             id = sample_code[0:9]
@@ -73,6 +100,7 @@ class ImageRegistrationProcess:
         return ds_id
 
     def registerImageFileInOmero(self, file_path, dataset_id):
+        
         cmd_list = list(self._init_cmd_list)
         cmd_list.append( "python backendinterface.py -f " + file_path + " -d " + str(dataset_id) )
 
@@ -84,6 +112,9 @@ class ImageRegistrationProcess:
         out, err = process.communicate( commands )
 
         id_list = str(out).split()
+        for img_id in id_list:
+            if not img_id.isdigit():
+                return []
 
         return id_list
         
@@ -91,16 +122,16 @@ class ImageRegistrationProcess:
     def triggerOMETiffConversion(self):
         pass
 
-    #ToDo Check if Metadata file is provided as was suggested in test.tsv provided by LK
-    def extractMetadataFromTSV(self, tsvFilePath):
+    #ToDo Check if Metadata file is provided as defined
+    def extractMetadataFromTSV(self, tsv_file_path):
         tsvFileMap = {}
         try:
-            with open(tsvFilePath) as tsvfile:
+            with open(tsv_file_path) as tsvfile:
                 reader = csv.DictReader(tsvfile, delimiter='\t', strict=True)
                 for row in reader:
                     tsvFileMap.update(row)
         except IOError:
-            print "Error: No file found at provided filepath " + tsvFilePath
+            print "Error: No file found at provided filepath " + tsv_file_path
         except csv.Error as e:
             print 'Could not gather the Metadata from TSVfile %s, in line %d: %s' % (tsvfile, reader.line_num, e)
 
@@ -108,6 +139,30 @@ class ImageRegistrationProcess:
 
     def registerExperimentDataInOpenBIS(self):
         pass
+
+    def registerOmeroKeyValuePairs(self, image_id, property_map):
+        """Registers the property map as key-value pairs in the OMERO server.
+        """
+        
+        cmd_list = list(self._init_cmd_list)
+
+        # string format: key1::value1//key2::value2//key3::value3//...
+        key_value_str = ""
+        for key in property_map.keys(): 
+            key_value_str = key_value_str + str(key) + "::" + str(property_map[key]) + "//"
+        key_value_str = key_value_str[:len(key_value_str)-2] #remove last two chars
+
+        cmd_list.append( "python backendinterface.py -i " + str(image_id) + " -a " + key_value_str )
+
+        commands = ""
+        for cmd in cmd_list:
+            commands = commands + cmd + "\n"
+
+        process = Popen( "/bin/bash", shell=False, universal_newlines=True, stdin=PIPE, stdout=PIPE, stderr=PIPE )
+        out, err = process.communicate( commands )
+
+
+        return 0
 
 
 class SampleCodeError(Exception):
@@ -120,4 +175,12 @@ class SampleCodeError(Exception):
     def test(self):
         pass
 
+class SampleNotFoundError(Exception):
+    
+    def __init__(self, sample_code, message):
+        self.sample_code = sample_code
+        self.message = message
+        super().__init__(self.message)
 
+    def test(self):
+        pass